Split out disk IO functions.
