+++ test/lib/ansible_test/_internal/metadata.py
@@ -11,7 +11,7 @@ from .util import (
     is_shippable,
 )
 
-from .util_common import (
+from .io import (
     write_json_file,
 )
