+++ test/lib/ansible_test/_internal/util_common.py
@@ -4,7 +4,6 @@ __metaclass__ = type
 
 import atexit
 import contextlib
-import json
 import os
 import shutil
 import sys
@@ -25,8 +24,12 @@ from .util import (
     raw_command,
     to_bytes,
     ANSIBLE_TEST_DATA_ROOT,
-    make_dirs,
     ApplicationError,
+)
+
+from .io import (
+    write_text_file,
+    write_json_file,
 )
 
 from .data import (
@@ -147,21 +150,6 @@ def write_text_test_results(category, name, content):  # type: (ResultType, str,
     """Write the given text content to the specified test results path, creating directories as needed."""
     path = os.path.join(category.path, name)
     write_text_file(path, content, create_directories=True)
-
-
-def write_json_file(path, content, create_directories=False):  # type: (str, t.Union[t.List[t.Any], t.Dict[str, t.Any]], bool) -> None
-    """Write the given json content to the specified path, optionally creating missing directories."""
-    text_content = json.dumps(content, sort_keys=True, indent=4) + '\n'
-    write_text_file(path, text_content, create_directories=create_directories)
-
-
-def write_text_file(path, content, create_directories=False):  # type: (str, str, bool) -> None
-    """Write the given text content to the specified path, optionally creating missing directories."""
-    if create_directories:
-        make_dirs(os.path.dirname(path))
-
-    with open(to_bytes(path), 'wb') as file:
-        file.write(to_bytes(content))
 
 
 def get_python_path(args, interpreter):
